model/requirements.txt
pymongo[srv]
fastapi
uvicorn
sentence-transformers
scikit-learn
python-dotenv
